requirements.txt
apache-airflow[postgres,celery,redis,password,ssh]